PHPUnit : Refactoring Tests (Use local resource)
