+++ .env.test
@@ -5,3 +5,4 @@ SYMFONY_DEPRECATIONS_HELPER=999999
 
 DATABASE_URL="mysql://root:password@127.0.0.1:3306/qanightlyresults?serverVersion=8.0.23&charset=utf8mb4"
 QANB_TOKEN='AZERTY'
+QANB_REPORT_PATH='./tests/resources/'
