+++ config/services.yaml
@@ -5,6 +5,7 @@
 # https://symfony.com/doc/current/best_practices.html#use-parameters-for-application-configuration
 parameters:
     QANB_GCPURL: 'https://storage.googleapis.com/prestashop-core-nightly/'
+    QANB_REPORT_PATH: '%env(default:QANB_GCPURL:QANB_REPORT_PATH)%'
     QANB_TOKEN: '%env(QANB_TOKEN)%'
 
 services:
@@ -13,7 +14,7 @@ services:
         autowire: true      # Automatically injects dependencies in your services.
         autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
         bind:
-            $nightlyGCPUrl: '%QANB_GCPURL%'
+            $nightlyReportPath: '%QANB_REPORT_PATH%'
             $nightlyToken: '%QANB_TOKEN%'
 
     # makes classes in src/ available to be used as services
