+++ tests/Controller/ImportControllerTest.php
@@ -6,17 +6,12 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 class ImportControllerTest extends WebTestCase
 {
-    private static string $date = '';
-
-    public static function setUpBeforeClass(): void
-    {
-        self::$date = date('Y-m-d', strtotime('today'));
-    }
+    private const DATE_RESOURCE = '2024-01-30';
 
     public function testReportOkAutoupgrade(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/hook/reports/import?filename=autoupgrade_' . self::$date . '-develop.json&token=AZERTY&campaign=autoupgrade&platform=cli');
+        $client->request('GET', '/hook/reports/import?filename=autoupgrade_' . self::DATE_RESOURCE . '-develop.json&token=AZERTY&campaign=autoupgrade&platform=cli');
 
         $response = $client->getResponse();
 
@@ -67,7 +62,7 @@ class ImportControllerTest extends WebTestCase
     public function testReportWithParameterFilename(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/hook/reports/import?filename=' . self::$date . '-develop.json');
+        $client->request('GET', '/hook/reports/import?filename=' . self::DATE_RESOURCE . '-develop.json');
         $response = $client->getResponse();
 
         $this->assertEquals(400, $response->getStatusCode());
@@ -83,7 +78,7 @@ class ImportControllerTest extends WebTestCase
     public function testReportWithParameterFilenameAndBakToken(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/hook/reports/import?filename=' . self::$date . '-develop.json&token=BAD');
+        $client->request('GET', '/hook/reports/import?filename=' . self::DATE_RESOURCE . '-develop.json&token=BAD');
         $response = $client->getResponse();
 
         $this->assertEquals(401, $response->getStatusCode());
@@ -99,7 +94,7 @@ class ImportControllerTest extends WebTestCase
     public function testReportWithNoVersionInFilename(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/hook/reports/import?filename=' . self::$date . '.json&token=AZERTY');
+        $client->request('GET', '/hook/reports/import?filename=' . self::DATE_RESOURCE . '.json&token=AZERTY');
         $response = $client->getResponse();
 
         $this->assertEquals(400, $response->getStatusCode());
@@ -115,7 +110,7 @@ class ImportControllerTest extends WebTestCase
     public function testReportWithBadVersionInFilename(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/hook/reports/import?filename=' . self::$date . '-.json&token=AZERTY');
+        $client->request('GET', '/hook/reports/import?filename=' . self::DATE_RESOURCE . '-.json&token=AZERTY');
         $response = $client->getResponse();
 
         $this->assertEquals(400, $response->getStatusCode());
@@ -125,13 +120,13 @@ class ImportControllerTest extends WebTestCase
         $content = $response->getContent();
         $content = json_decode($content, true);
         $this->assertArrayHasKey('message', $content);
-        $this->assertEquals('Version found not correct () from filename ' . self::$date . '-.json', $content['message']);
+        $this->assertEquals('Version found not correct () from filename ' . self::DATE_RESOURCE . '-.json', $content['message']);
     }
 
     public function testReportWithNotExistingFilename(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/hook/reports/import?filename=' . self::$date . '-truc.json&token=AZERTY');
+        $client->request('GET', '/hook/reports/import?filename=' . self::DATE_RESOURCE . '-truc.json&token=AZERTY');
         $response = $client->getResponse();
 
         $this->assertEquals(400, $response->getStatusCode());
@@ -147,7 +142,7 @@ class ImportControllerTest extends WebTestCase
     public function testReportOk(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/hook/reports/import?filename=' . self::$date . '-develop.json&token=AZERTY');
+        $client->request('GET', '/hook/reports/import?filename=' . self::DATE_RESOURCE . '-develop.json&token=AZERTY');
         $response = $client->getResponse();
 
         $this->assertEquals(200, $response->getStatusCode());
@@ -165,7 +160,7 @@ class ImportControllerTest extends WebTestCase
     public function testReportAlreadyExisting(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/hook/reports/import?filename=' . self::$date . '-develop.json&token=AZERTY');
+        $client->request('GET', '/hook/reports/import?filename=' . self::DATE_RESOURCE . '-develop.json&token=AZERTY');
         $response = $client->getResponse();
 
         $this->assertEquals(403, $response->getStatusCode());
@@ -175,6 +170,6 @@ class ImportControllerTest extends WebTestCase
         $content = $response->getContent();
         $content = json_decode($content, true);
         $this->assertArrayHasKey('message', $content);
-        $this->assertEquals('A similar entry was found (criteria: version develop, platform chromium, campaign functional, date ' . self::$date . ').', $content['message']);
+        $this->assertEquals('A similar entry was found (criteria: version develop, platform chromium, campaign functional, date ' . self::DATE_RESOURCE . ').', $content['message']);
     }
 }
