+++ tests/resources/reports/autoupgrade_2024-01-25-develop.json
@@ -1,0 +1,1 @@
+{"stats":{"suites":1,"tests":2,"passes":0,"pending":0,"failures":2,"testsRegistered":2,"other":0,"hasOther":0,"skipped":0,"hasSkipped":0,"start":"2024-01-25T00:10:12.000+00:00","end":"2024-01-25T00:12:11.000+00:00","duration":121000},"results":[{"uuid":"65b1a80c482ae","title":"Upgrade to branch develop","fullFile":"","file":"","tests":[],"suites":[{"uuid":"65b1a80c482ae","title":"[master] Upgrade to branch develop","file":"","tests":[{"title":"[master] Upgrade from 8.0.4 to develop","fullTitle":"[master] Upgrade from 8.0.4 to develop","timedOut":false,"duration":121000,"state":"failed","pass":false,"fail":true,"pending":false,"context":"{\"title\": \"testIdentifier\",\"value\": \"[master] Upgrade from 8.0.4 to develop\"}","uuid":"65b1a80c4827d","parentUUID":"65b1a80c482ae","isHook":false,"skipped":false}],"suites":[],"passes":[],"failures":["65b1a80c4827d"],"pending":[],"skipped":[],"duration":121000,"root":false,"rootEmpty":false}],"passes":[],"failures":[],"pending":[],"skipped":[],"duration":121000,"root":true,"rootEmpty":true},{"uuid":"65b1a80c482ba","title":"Upgrade to branch develop","fullFile":"","file":"","tests":[],"suites":[{"uuid":"65b1a80c482ba","title":"[dev] Upgrade to branch develop","file":"","tests":[{"title":"[dev] Upgrade from 8.0.4 to develop","fullTitle":"[dev] Upgrade from 8.0.4 to develop","timedOut":false,"duration":118000,"state":"failed","pass":false,"fail":true,"pending":false,"context":"{\"title\": \"testIdentifier\",\"value\": \"[dev] Upgrade from 8.0.4 to develop\"}","uuid":"65b1a80c482ab","parentUUID":"65b1a80c482ba","isHook":false,"skipped":false}],"suites":[],"passes":[],"failures":["65b1a80c482ab"],"pending":[],"skipped":[],"duration":118000,"root":false,"rootEmpty":false}],"passes":[],"failures":[],"pending":[],"skipped":[],"duration":121000,"root":true,"rootEmpty":true}]}
