+++ tests/resources/reports/autoupgrade_2024-01-30-develop.json
@@ -1,0 +1,1 @@
+{"stats":{"suites":1,"tests":2,"passes":0,"pending":0,"failures":2,"testsRegistered":2,"other":0,"hasOther":0,"skipped":0,"hasSkipped":0,"start":"2024-01-30T00:09:47.000+00:00","end":"2024-01-30T00:11:51.000+00:00","duration":124000},"results":[{"uuid":"65b892c16840d","title":"Upgrade to branch develop","fullFile":"","file":"","tests":[],"suites":[{"uuid":"65b892c16840d","title":"[dev] Upgrade to branch develop","file":"","tests":[{"title":"[dev] Upgrade from 8.0.4 to develop","fullTitle":"[dev] Upgrade from 8.0.4 to develop","timedOut":false,"duration":124000,"state":"failed","pass":false,"fail":true,"pending":false,"context":"{\"title\": \"testIdentifier\",\"value\": \"[dev] Upgrade from 8.0.4 to develop\"}","uuid":"65b892c1683e5","parentUUID":"65b892c16840d","isHook":false,"skipped":false}],"suites":[],"passes":[],"failures":["65b892c1683e5"],"pending":[],"skipped":[],"duration":124000,"root":false,"rootEmpty":false}],"passes":[],"failures":[],"pending":[],"skipped":[],"duration":124000,"root":true,"rootEmpty":true},{"uuid":"65b892c168416","title":"Upgrade to branch develop","fullFile":"","file":"","tests":[],"suites":[{"uuid":"65b892c168416","title":"[master] Upgrade to branch develop","file":"","tests":[{"title":"[master] Upgrade from 8.0.4 to develop","fullTitle":"[master] Upgrade from 8.0.4 to develop","timedOut":false,"duration":21607000,"state":"failed","pass":false,"fail":true,"pending":false,"context":"{\"title\": \"testIdentifier\",\"value\": \"[master] Upgrade from 8.0.4 to develop\"}","uuid":"65b892c16840a","parentUUID":"65b892c168416","isHook":false,"skipped":false}],"suites":[],"passes":[],"failures":["65b892c16840a"],"pending":[],"skipped":[],"duration":21607000,"root":false,"rootEmpty":false}],"passes":[],"failures":[],"pending":[],"skipped":[],"duration":124000,"root":true,"rootEmpty":true}]}
